Created avatarHandler.js
This file provide a client side handler for getting avatar image and uploading image. previewFile() get an image file with a file picker/reader, transform it into an url and display the image in a <img>. sendForm() interrupt the image upload to make client checking on the image url to send and may refuse it.